third_party/WebKit/LayoutTests/fast/css/background-image-leak-expected.txt

PASS The difference of counter "numberOfLiveDocuments" before and after the cycle is under the threshold of 0.
PASS successfullyParsed is true

TEST COMPLETE
